+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/data/set/filter/CommerceAccountAutocompleteFDSFilter.java
@@ -27,16 +27,17 @@ import org.osgi.service.component.annotations.Component;
 	property = "frontend.data.set.name=" + CommerceOrderFDSNames.ALL_ORDERS,
 	service = FDSFilter.class
 )
-public class CommerceChannelFDSFilter extends BaseAutocompleteFDSFilter {
+public class CommerceAccountAutocompleteFDSFilter
+	extends BaseAutocompleteFDSFilter {
 
 	@Override
 	public String getAPIURL() {
-		return "/o/headless-commerce-admin-channel/v1.0/channels?sort=name:asc";
+		return "/o/headless-commerce-admin-account/v1.0/accounts?sort=name:asc";
 	}
 
 	@Override
 	public String getId() {
-		return "channelId";
+		return "accountId";
 	}
 
 	@Override
@@ -51,12 +52,12 @@ public class CommerceChannelFDSFilter extends BaseAutocompleteFDSFilter {
 
 	@Override
 	public String getLabel() {
-		return "channel";
+		return "account";
 	}
 
 	@Override
 	public boolean isMultipleSelection() {
-		return false;
+		return true;
 	}
 
 }
